requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyYAML==6.0.2
python-multipart==0.0.20
pytest==8.4.1
